chore: dont export unneeded types
